Did some important and necessary changes in your website
There was an unavoidable mistake in your index.html file. 
When I was looking at your website in a browser it was showing just the index.html file there was no CSS applied to it. So, when I checked the index.html file, I saw you had linked the CSS file as href="css/style.css" and when I saw your folder there was no CSS folder so I changed it to  href="style.css" and then I was able to see your proper website. 
Also the same applies to the JS file you included the main folder JS which was not present due to which JS file was not working. 
I have added a website icon you can see at the top of the browser to upgrade your website. You can change the image as you wish.
You have made an awesome website. Hope this suggestion helped you : )